修复一些docx转html的BUG

+++ thirdparties-extension/fr.opensagres.poi.xwpf.converter.core/src/main/java/fr/opensagres/poi/xwpf/converter/core/utils/DxaUtil.java
@@ -60,7 +60,7 @@ public class DxaUtil
 
     public static float dxa2points( BigInteger dxa )
     {
-        return dxa.intValue() / 20;
+        return (float) (dxa.doubleValue() / 20);
     }
 
     public static float dxa2inch( float dxa )
